+++ R/CovarianceM.R
@@ -21,41 +21,37 @@
 #' @author Katherine L. Valeriano, Victor H. Lachos and Larissa A. Matos
 #'
 #' @examples
-#' # Initial parameter values
-#' phi <- 5;     rho <- 0.45
-#' tau2 <- 0.80; sigma2 <- 2
-#' # Simulating data
-#' n1 <- 10   # Number of spatial locations
-#' n2 <- 5    # Number of temporal index
 #' set.seed(1000)
-#' x.co <- round(runif(n1,0,10),5)  # X coordinate
-#' y.co <- round(runif(n1,0,10),5)  # Y coordinate
-#' coord <- cbind(x.co,y.co)        # Cartesian coordinates without repetitions
-#' time <- as.matrix(seq(1,n2))     # Time index without repetitions
+#' # Parameter values
+#' phi  <- 5
+#' rho  <- 0.45
+#' tau2 <- 0.80
+#' sigma2 <- 2
+#'
+#' # Coordinates and time points
+#' coords <- matrix(runif(20, 0, 10), ncol=2) # Cartesian coordinates without repetitions
+#' time   <- as.matrix(1:5)   # Time index without repetitions
+#' Ms <- as.matrix(dist(coords)) # Spatial distances
+#' Mt <- as.matrix(dist(time))   # Temporal distances
+#'
 #' # Covariance matrix
-#' Ms <- as.matrix(dist(coord))     # Spatial distances
-#' Mt <- as.matrix(dist(time))      # Temporal distances
-#' Cov <- CovarianceM(phi,rho,tau2,sigma2,distSpa=Ms,disTemp=Mt,kappa=0,type.S="exponential")
+#' Cov <- CovarianceM(phi, rho, tau2, sigma2, distSpa=Ms, disTemp=Mt,
+#'                    kappa=0, type.S="exponential")
 
 CovarianceM = function(phi, rho, tau2, sigma2, distSpa, disTemp, kappa, type.S){
 
   if (phi <= 0){stop("The spatial parameter can not be negative or equal to zero")}
-
   if (rho>1 | rho<(-1)){stop("The time scaling parameter can not be >1 or < -1")}
-
   if (tau2 < 0){stop("The nugget effect can not be negative")}
-
   if (sigma2 < 0){stop("The partial sill can not be negative")}
 
   if (ncol(distSpa) != nrow(distSpa)) stop("Spatial distance matrix must be specified")
-
   if (ncol(disTemp) != nrow(disTemp)) stop("Temporal distance matrix must be specified")
 
-  if (type.S!="matern" & type.S !="gaussian" & type.S != "spherical" & type.S != "pow.exp" & type.S != "exponential"){
+  if (!type.S%in%c("matern","gaussian","spherical","pow.exp","exponential")){
     stop('type.S should be one of matern, gaussian, spherical, pow.exp, exponential')}
 
   if (type.S=="pow.exp" & (kappa > 2| kappa<=0)) stop("kappa must be a real in (0,2]")
-
   if (type.S=="matern" & kappa <= 0) stop("kappa must be a real number in (0,Inf)")
 
   #---------------------------------------------------------------------#